+++ src/impl/x86_64/print.cpp
@@ -118,7 +118,7 @@ void TerminalPrinter::print_str(const char *str, ...)
 				}
 				case 'x':
 				{
-					int a = va_arg(ap, int);
+					uint64_t a = va_arg(ap, uint64_t);
 					print_str("0x");
 					print_str(hexToString(a));
 					i+=1;
@@ -126,7 +126,7 @@ void TerminalPrinter::print_str(const char *str, ...)
 				}
 				case 'X':
 				{
-					int a = va_arg(ap, int);
+					uint64_t a = va_arg(ap, uint64_t);
 					print_str("0x");
 					print_str(hexToString(a,true));
 					i+=1;
